+++ packages/client/shared/tiptap/extensions/InsightsBlockBase.ts
@@ -12,10 +12,6 @@ export const InsightsBlockBase = Node.create({
 
   content: 'block*',
 
-  draggable: true,
-
-  selectable: true,
-
   inline: false,
 
   addAttributes() {